retry in next batch, refactor
