+++ ghost/core/core/server/services/welcome-emails/jobs/process-outbox.js
@@ -1,20 +1,47 @@
 const {parentPort} = require('worker_threads');
 const logging = require('@tryghost/logging');
-const errors = require('@tryghost/errors');
 const MemberCreatedEvent = require('../../../../shared/events/MemberCreatedEvent');
 const {OUTBOX_STATUSES} = require('../../../models/outbox');
+const {MESSAGES, MAX_ENTRIES_PER_JOB, BATCH_SIZE} = require('./lib/constants');
+const processEntries = require('./lib/process-entries');
 
-const BATCH_SIZE = 100;
-const MAX_ENTRIES_PER_JOB = 1000;
-const MAX_RETRIES = 3;
-const SIMULATE_FAILURE_RATE = 0.3;
-
-const MESSAGES = {
-    CANCELLED: 'Outbox processing cancelled',
-    NO_ENTRIES: 'No pending outbox entries to process',
-    SIMULATION_MODE: `Welcome email processor running in simulation mode (${SIMULATE_FAILURE_RATE * 100}% failure rate)`,
-    SIMULATED_FAILURE: 'Simulated random failure for testing retry logic'
-};
+/**
+ * Fetches pending outbox entries with exponential backoff and locks them for processing
+ * @param {Object} db - Database connection
+ * @param {number} batchSize - Maximum number of entries to fetch
+ * @param {string} jobStartISO - ISO-formatted timestamp when the job started (YYYY-MM-DD HH:mm:ss)
+ * @returns {Promise<Array>} Array of outbox entries marked as PROCESSING
+ */
+async function fetchPendingEntries(db, batchSize, jobStartISO) {
+    let entries = [];
+    await db.knex.transaction(async (trx) => {
+        const query = trx('outbox')
+            .where('event_type', MemberCreatedEvent.name)
+            .where('status', OUTBOX_STATUSES.PENDING)
+            .where(function () {
+                this.whereNull('last_retry_at')
+                    .orWhere('last_retry_at', '<', jobStartISO);
+            });
+        
+        entries = await query
+            .orderBy('created_at', 'asc')
+            .limit(batchSize)
+            .forUpdate()
+            .select('*');
+        
+        if (entries.length > 0) {
+            const ids = entries.map(e => e.id);
+            await trx('outbox')
+                .whereIn('id', ids)
+                .update({
+                    status: OUTBOX_STATUSES.PROCESSING,
+                    updated_at: db.knex.raw('CURRENT_TIMESTAMP')
+                });
+        }
+    });
+    
+    return entries;
+}
 
 /**
  * Sends a status message to the parent worker thread
@@ -55,120 +82,6 @@ function cancel() {
     }
 }
 
-/**
- * Fetches pending outbox entries with exponential backoff and locks them for processing
- * @param {Object} db - Database connection
- * @param {number} batchSize - Maximum number of entries to fetch
- * @returns {Promise<Array>} Array of outbox entries marked as PROCESSING
- */
-async function fetchPendingEntries(db, batchSize) {
-    let entries = [];
-    await db.knex.transaction(async (trx) => {
-        const query = trx('outbox')
-            .where('event_type', MemberCreatedEvent.name)
-            .where('status', OUTBOX_STATUSES.PENDING);
-        
-        entries = await query
-            .orderBy('created_at', 'asc')
-            .limit(batchSize)
-            .forUpdate()
-            .select('*');
-        
-        if (entries.length > 0) {
-            const ids = entries.map(e => e.id);
-            await trx('outbox')
-                .whereIn('id', ids)
-                .update({
-                    status: OUTBOX_STATUSES.PROCESSING,
-                    updated_at: db.knex.raw('CURRENT_TIMESTAMP')
-                });
-        }
-    });
-    
-    return entries;
-}
-
-/**
- * Deletes a successfully processed outbox entry
- * @param {Object} db - Database connection
- * @param {string} entryId - ID of the entry to delete
- */
-async function deleteProcessedEntry(db, entryId) {
-    await db.knex('outbox')
-        .where('id', entryId)
-        .delete();
-}
-
-/**
- * Updates a failed entry with incremented retry count and new status
- * @param {Object} options - Update options
- * @param {Object} options.db - Database connection
- * @param {string} options.entryId - ID of the entry to update
- * @param {number} options.retryCount - Current retry count
- * @param {string} options.errorMessage - Error message to store
- */
-async function updateFailedEntry({db, entryId, retryCount, errorMessage}) {
-    const newRetryCount = retryCount + 1;
-    const newStatus = newRetryCount >= MAX_RETRIES ? OUTBOX_STATUSES.FAILED : OUTBOX_STATUSES.PENDING;
-
-    await db.knex('outbox')
-        .where('id', entryId)
-        .update({
-            status: newStatus,
-            retry_count: newRetryCount,
-            last_retry_at: db.knex.raw('CURRENT_TIMESTAMP'),
-            message: errorMessage.substring(0, 2000),
-            updated_at: db.knex.raw('CURRENT_TIMESTAMP')
-        });
-}
-
-/**
- * Determines whether to simulate a failure for testing retry logic
- * @returns {boolean} True if failure should be simulated (non-production only)
- */
-function shouldSimulateFailure() {
-    return process.env.NODE_ENV !== 'production' && Math.random() < SIMULATE_FAILURE_RATE;
-}
-
-/**
- * Sends a welcome email to a new member (currently logs only)
- * @param {Object} payload - Member data containing name and email
- */
-async function sendWelcomeEmail(payload) {
-    logging.info(`[WELCOME-EMAIL] Welcome email sent to ${payload.name} at ${payload.email}`);
-}
-
-/**
- * Processes a single outbox entry by sending welcome email and managing retry logic
- * @param {Object} db - Database connection
- * @param {Object} entry - Outbox entry to process
- * @returns {Promise<Object>} Result object with success boolean
- */
-async function processEntry(db, entry) {
-    let payload;
-    try {
-        payload = JSON.parse(entry.payload);
-
-        if (shouldSimulateFailure()) {
-            throw new errors.InternalServerError({
-                message: MESSAGES.SIMULATED_FAILURE
-            });
-        }
-
-        await sendWelcomeEmail(payload);
-        await deleteProcessedEntry(db, entry.id);
-
-        return {success: true};
-    } catch (err) {
-        await updateFailedEntry({db, entryId: entry.id, retryCount: entry.retry_count, errorMessage: err.message});
-
-        const memberInfo = payload ? `${payload.name} (${payload.email})` : 'unknown member';
-        logging.error(`[WELCOME-EMAIL] Failed to send to ${memberInfo}: ${err.message}`);
-
-        return {success: false};
-    }
-}
-
 if (parentPort) {
     parentPort.once('message', (message) => {
         if (message === 'cancel') {
@@ -177,43 +90,11 @@ if (parentPort) {
     });
 }
 
-/**
- * Formats the job completion message with stats
- * @param {Object} options - Completion stats
- * @param {number} options.processed - Number of successfully processed entries
- * @param {number} options.failed - Number of failed entries
- * @param {number} options.durationMs - Total processing time in milliseconds
- * @returns {string} Formatted completion message
- */
-function formatCompletionMessage({processed, failed, durationMs}) {
-    return `Processed ${processed} outbox entries, ${failed} failed in ${(durationMs / 1000).toFixed(2)}s`;
-}
-
-/**
- * Processes all entries in a batch sequentially
- * @param {Object} db - Database connection
- * @param {Array} entries - Array of outbox entries to process
- * @returns {Promise<Object>} Object with processed and failed counts
- */
-async function processEntries(db, entries) {
-    let processed = 0;
-    let failed = 0;
-
-    for (const entry of entries) {
-        const result = await processEntry(db, entry);
-        if (result.success) {
-            processed += 1;
-        } else {
-            failed += 1;
-        }
-    }
-
-    return {processed, failed};
-}
-
-(async () => {
-    const startTime = Date.now();
+async function processOutbox() {
     const db = require('../../../data/db');
+    
+    const jobStartMs = Date.now();
+    const jobStartISO = new Date(jobStartMs).toISOString().slice(0, 19).replace('T', ' ');
 
     if (process.env.NODE_ENV !== 'production') {
         logging.info(MESSAGES.SIMULATION_MODE);
@@ -221,9 +102,16 @@ async function processEntries(db, entries) {
 
     let totalProcessed = 0;
     let totalFailed = 0;
-    let entries = await fetchPendingEntries(db, BATCH_SIZE);
 
-    while (entries.length > 0 && totalProcessed + totalFailed < MAX_ENTRIES_PER_JOB) {
+    while (totalProcessed + totalFailed < MAX_ENTRIES_PER_JOB) {
+        const remainingCapacity = MAX_ENTRIES_PER_JOB - (totalProcessed + totalFailed);
+        const fetchSize = Math.min(BATCH_SIZE, remainingCapacity);
+        
+        const entries = await fetchPendingEntries(db, fetchSize, jobStartISO);
+        if (entries.length === 0) {
+            break;
+        }
+
         const batchStartMs = Date.now();
         const {processed, failed} = await processEntries(db, entries);
         const batchDurationMs = Date.now() - batchStartMs;
@@ -233,15 +121,15 @@ async function processEntries(db, entries) {
         totalFailed += failed;
 
         logging.info(`[WELCOME-EMAIL] Batch complete: ${processed} processed, ${failed} failed in ${(batchDurationMs / 1000).toFixed(2)}s (${batchRate} entries/sec)`);
-
-        entries = await fetchPendingEntries(db, BATCH_SIZE);
     }
 
-    const durationMs = Date.now() - startTime;
+    const durationMs = Date.now() - jobStartMs;
 
     if (totalProcessed + totalFailed === 0) {
         return completeJob(MESSAGES.NO_ENTRIES);
     }
 
-    completeJob(formatCompletionMessage({processed: totalProcessed, failed: totalFailed, durationMs}));
-})();
+    completeJob(`[WELCOME-EMAIL] Job complete: Processed ${totalProcessed} outbox entries, ${totalFailed} failed in ${(durationMs / 1000).toFixed(2)}s`);
+}
+
+processOutbox();
